Began Kit_API setup
