+++ src/classes/Kit.java
@@ -1,0 +1,17 @@
+package classes;
+
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.Inventory;
+import utilities.PVPUtils;
+
+import java.util.HashSet;
+
+public interface Kit {
+
+    String kitName();
+
+    String getDisplay();
+
+    Inventory getPlayerArmor();
+    void setArmor();
+}
