+++ src/classes/KitValues.java
@@ -1,0 +1,20 @@
+package classes;
+
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+
+/**
+ * Created by Signifies on 10/10/2018 - 23:49.
+ */
+public enum KitValues
+{
+
+    NO_DEBUFF(),
+    DEBUF(),
+    ARCHER(),
+    COMBO(),
+    GAPPLE(),
+    SUMO(),
+
+
+}
