+++ src/notes
@@ -1,0 +1,1 @@
+-Check kBattle code see if I have arena and duel system implemented.
